+++ delphin/mrs/dmrx.py
@@ -14,7 +14,6 @@ import xml.etree.ElementTree as etree
 from delphin.mrs import (Dmrs, Node, Link, Pred, Lnk)
 from delphin.mrs.components import (nodes, links)
 from delphin.mrs.config import QUANTIFIER_POS
-from delphin.mrs.util import etree_tostring
 
 ##############################################################################
 ##############################################################################
@@ -220,11 +219,11 @@ def serialize(ms, properties=True, encoding='unicode', pretty_print=False):
     # for now, pretty_print=True is the same as pretty_print='LKB'
     if pretty_print in ('LKB', 'lkb', 'Lkb', True):
         lkb_pprint_re = re.compile(r'(<dmrs[^>]+>|</node>|</link>|</dmrs>)')
-        string = str(etree_tostring(e, encoding=encoding))
+        string = str(etree.tostring(e, encoding=encoding))
         return lkb_pprint_re.sub(r'\1\n', string)
     # pretty_print is only lxml. Look into tostringlist, maybe?
     # return etree.tostring(e, pretty_print=pretty_print, encoding='unicode')
-    return etree_tostring(e, encoding=encoding)
+    return etree.tostring(e, encoding=encoding)
 
 
 def _encode_dmrs(m, properties):